moved general harvest methods into map_env, pep8
